backend/python-ocr-service/requirements.txt
pytesseract
Pillow
Flask
